refactor caching
