+++ src/core.rs
@@ -2,31 +2,13 @@
 
 use crate::{
     db_kind::DbKind,
-    db_value::{ColumnMap, DbParams, DbRow, DbRows, DbValue, IntoDbParams, IntoDbRows},
-    memory::{
-        DEFAULT_MEMORY_QUERY_CACHE_SIZE, MemoryQueryCacheKey, MemoryQueryCacheValue,
-        clear_memory_query_cache, exists_in_meta_cache, get_memory_query_cache,
-        get_memory_table_cache, get_meta_cache,
-    },
-    params,
-    parse::{get_accessed_tables, get_affected_tables, get_view_tables, validate_table_name},
+    db_value::{ColumnMap, DbParams, DbRows, IntoDbParams, IntoDbRows},
+    memory::{exists_in_meta_cache, get_meta_cache},
 };
 
 use async_trait::async_trait;
 use serde::{de, ser};
-use std::{
-    collections::HashSet,
-    fmt::Display,
-    future::Future,
-    str::FromStr,
-    time::{SystemTime, UNIX_EPOCH},
-};
-
-/// The name of the query cache table.
-pub static QUERY_CACHE_TABLE: &str = "rltbl_db_query_cache";
-
-/// The name of the table cache table.
-pub static TABLE_CACHE_TABLE: &str = "rltbl_db_table_cache";
+use std::{fmt::Display, future::Future};
 
 /// Defines the various error types that can be generated by rltbl_db.
 #[derive(Clone, Debug, PartialEq)]
@@ -82,417 +64,10 @@ impl std::fmt::Display for DbError {
     }
 }
 
-/// Strategy to use when caching query results
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum CachingStrategy {
-    /// No Caching.
-    None,
-    /// Truncate the entire cache when it is dirty.
-    TruncateAll,
-    /// Truncate entries only for edited tables when the cache is dirty.
-    Truncate,
-    /// Truncate cache entries, for edited tables only, automatically whenever tables are edited.
-    Trigger,
-    /// Similar to Truncate, but use an in-memory cache.
-    Memory(usize),
-}
-
-impl FromStr for CachingStrategy {
-    type Err = DbError;
-
-    fn from_str(strategy: &str) -> Result<Self, DbError> {
-        match strategy.to_lowercase().as_str() {
-            "none" => Ok(CachingStrategy::None),
-            "truncate_all" => Ok(CachingStrategy::TruncateAll),
-            "truncate" => Ok(CachingStrategy::Truncate),
-            "trigger" => Ok(CachingStrategy::Trigger),
-            strategy if strategy.starts_with("memory") => {
-                let elems = strategy.split(":").collect::<Vec<_>>();
-                let cache_size = {
-                    if elems.len() < 2 {
-                        DEFAULT_MEMORY_QUERY_CACHE_SIZE
-                    } else {
-                        let cache_size = elems[1];
-                        match cache_size.parse::<usize>() {
-                            Ok(0) => DEFAULT_MEMORY_QUERY_CACHE_SIZE,
-                            Ok(size) => size,
-                            Err(err) => return Err(DbError::InputError(format!(
-                                "Error parsing memory cache size specification: '{cache_size}': \
-                                 {err}"
-                            ))
-                            .into()),
-                        }
-                    }
-                };
-                Ok(CachingStrategy::Memory(cache_size))
-            }
-            _ => {
-                return Err(
-                    DbError::InputError(format!("Unrecognized strategy: {strategy}")).into(),
-                );
-            }
-        }
-    }
-}
-
-impl Display for CachingStrategy {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            CachingStrategy::None => write!(f, "none"),
-            CachingStrategy::TruncateAll => write!(f, "truncate_all"),
-            CachingStrategy::Truncate => write!(f, "truncate"),
-            CachingStrategy::Trigger => write!(f, "trigger"),
-            CachingStrategy::Memory(size) => write!(f, "memory:{size}"),
-        }
-    }
-}
-
 #[async_trait]
 pub trait DbQuery {
     /// Get the kind of SQL database. See [DbKind] for the supported database kinds.
     fn kind(&self) -> DbKind;
-
-    /// Set the caching strategy.
-    fn set_caching_strategy(&mut self, strategy: &CachingStrategy);
-
-    /// Get the current caching strategy.
-    fn get_caching_strategy(&self) -> CachingStrategy;
-
-    /// When turned on, and the current [CachingStrategy] is not [None](CachingStrategy::None),
-    /// SQL commands executed through the API will be automatically checked to see if they
-    /// involve edits and/or drops of database tables. If they do then the cache will be
-    /// automatically updated in accordance with the current [CachingStrategy].
-    /// Note that setting this flag does not imply that the results of queries should be
-    /// cached. Setting this flag only means that the current contents, if any, of the cache
-    /// table should be kept up to date whenever the data in the database is edited
-    /// via one of the query_* or execute() methods in [DbQuery]. To add new content to the cache
-    /// that can be later be reused you must explicitly use the [DbQuery::cache()] method.
-    /// To explicitly skip the housekeeping implied by setting the cache-aware-query flag, even
-    /// when it is set to on, use [DbQuery::execute_no_cache_clean()] or
-    /// [DbQuery::query_no_cache_clean()].
-    fn set_cache_aware_query(&mut self, value: bool);
-
-    /// Returns true if the cache-aware-query option is currently on.
-    fn get_cache_aware_query(&self) -> bool;
-
-    /// Ensure that the query cache table and the table cache table exist (see
-    /// [QUERY_CACHE_TABLE] and [TABLE_CACHE_TABLE]).
-    async fn ensure_cache_tables_exist(&self) -> Result<(), DbError> {
-        if !exists_in_meta_cache(QUERY_CACHE_TABLE)? || !exists_in_meta_cache(TABLE_CACHE_TABLE)? {
-            for special_table in [QUERY_CACHE_TABLE, TABLE_CACHE_TABLE] {
-                let sql = match special_table {
-                    table if table == QUERY_CACHE_TABLE => {
-                        self.kind().create_query_cache_table_sql()
-                    }
-                    table if table == TABLE_CACHE_TABLE => {
-                        self.kind().create_table_cache_table_sql()
-                    }
-                    _ => unreachable!(),
-                };
-                match self.execute_no_cache_clean(&sql, ()).await {
-                    Ok(_) => (),
-                    Err(_) => {
-                        // Since we are not using transactions, a race condition could occur in
-                        // which two or more threads are trying to create the cache at the same
-                        // time, triggering a primary key violation in the metadata table. So if
-                        // there is an error creating the cache table we just check that it exists
-                        // and if it does we assume that all is ok.
-                        match self.table_exists(special_table).await? {
-                            false => {
-                                return Err(DbError::DatabaseError(format!(
-                                    "The cache table '{special_table}' could not be created"
-                                )));
-                            }
-                            true => (),
-                        }
-                    }
-                };
-                let mut cache = get_meta_cache()?;
-                cache.insert(special_table.to_string());
-            }
-        }
-        Ok(())
-    }
-
-    /// Ensure that caching triggers exist for the given table. Note that this function calls
-    /// [DbQuery::ensure_cache_tables_exist()] implicitly.
-    async fn ensure_caching_triggers_exist_for_table(&self, table: &str) -> Result<(), DbError> {
-        let table_triggers_name = format!("{table}_triggers");
-        if !exists_in_meta_cache(&table_triggers_name)? {
-            self.ensure_cache_tables_exist().await?;
-            let sql = self
-                .kind()
-                .create_table_caching_triggers_for_table_sql(&table)?
-                .join(";\n");
-            self.execute_batch(&sql).await?;
-
-            // Indicate that triggers exist for `table` in the meta-cache:
-            let mut cache = get_meta_cache()?;
-            cache.insert(table_triggers_name);
-        }
-        Ok(())
-    }
-
-    /// Ensure that caching triggers exist for the source tables of the given view. Note that
-    /// this function calls [DbQuery::ensure_cache_tables_exist()] implicitly.
-    async fn ensure_caching_triggers_exist_for_view(&self, view: &str) -> Result<(), DbError> {
-        let view_triggers_name = format!("{view}_triggers");
-        if !exists_in_meta_cache(&view_triggers_name)? {
-            self.ensure_cache_tables_exist().await?;
-            let view_sql = self.get_view_sql(&view).await?;
-            let source_tables = get_view_tables(&view_sql)?;
-            for source_table in source_tables.iter() {
-                // Add a trigger to clean entries from the cache for the source table itself:
-                let sql = self
-                    .kind()
-                    .create_table_caching_triggers_for_table_sql(&source_table)?
-                    .join(";\n");
-                self.execute_batch(&sql).await?;
-                // Add a trigger to clean entries from the cache for the view:
-                let sql = self
-                    .kind()
-                    .create_table_caching_triggers_for_view_sql(&source_table, &view)?
-                    .join(";\n");
-                self.execute_batch(&sql).await?;
-                // Add an entry for the source table triggers to the metacache. If there is another
-                // entry for this source table it will be overwritten, which is desirable in
-                // case it was not previously known if the table was the source table for a view.
-                let mut cache = get_meta_cache()?;
-                let source_triggers_name = format!("{source_table}_triggers");
-                cache.insert(source_triggers_name);
-            }
-            let mut cache = get_meta_cache()?;
-            cache.insert(view_triggers_name);
-        }
-        Ok(())
-    }
-
-    /// Parse the given semi-colon-separated SQL commands and determine which tables will be
-    /// affected (either edited or dropped) by the commands, then ensure that there are no
-    /// entries for those tables in the cache in accordance with the current [CachingStrategy].
-    async fn clear_cache_for_affected_tables(&self, sql: &str) -> Result<(), DbError> {
-        if self.get_caching_strategy() != CachingStrategy::None {
-            let (edited_tables, dropped_tables): (Vec<_>, Vec<_>) = {
-                let (edited_tables, dropped_tables) = get_affected_tables(sql)?;
-                (
-                    edited_tables.into_iter().collect(),
-                    dropped_tables.into_iter().collect(),
-                )
-            };
-            if !edited_tables.is_empty() {
-                let edited_tables: Vec<_> = edited_tables.iter().map(|t| t.as_str()).collect();
-                self.clear_cache_for_edited_tables(&edited_tables).await?;
-            }
-            if !dropped_tables.is_empty() {
-                let dropped_tables: Vec<_> = dropped_tables.iter().map(|t| t.as_str()).collect();
-                self.clear_cache_for_dropped_tables(&dropped_tables).await?;
-            }
-        }
-        Ok(())
-    }
-
-    // Triggers cannot apply to DROP commands, only to INSERT, UPDATE, DELETE, or TRUNCATE.
-    // See https://www.postgresql.org/docs/current/sql-createtrigger.html and
-    // https://sqlite.org/lang_createtrigger.html. Note that PostgreSQL has the concept
-    // of an "event trigger":
-    // https://www.pgtutorial.com/postgresql-tutorial/postgresql-event-triggers/ which could
-    // be used, but SQLite has no such capability. To workaround this limitation, we
-    // define two clear_cache_() functions, one for edited tables, and one for dropped tables.
-    // In the case of a dropped table, unlike an edit, we cannot rely on the caching trigger,
-    // when we are using the [CachingStategy::Trigger] strategy, to automatically delete the
-    // entries from the cache for those tables, since those triggers will have beeen dropped
-    // along with the table.
-    // Although strictly speaking, PostgreSQL (which has event triggers) is not subject to this
-    // limitation, for simplicity we will not be creating a PostgreSQL event trigger and we will
-    // use both functions below for both database types.
-
-    /// Update the cache tables, for the given list of tables, using the current [CachingStrategy],
-    /// under the assumption that the tables in the given list have all just been edited (i.e.,
-    /// truncated, deleted from, inserted to, or updated).
-    async fn clear_cache_for_edited_tables(&self, tables: &[&str]) -> Result<(), DbError> {
-        match self.get_caching_strategy() {
-            CachingStrategy::None | CachingStrategy::Trigger => (),
-            CachingStrategy::TruncateAll => {
-                self.update_last_modified_times(tables).await?;
-                self.delete_query_cache_entries(&[]).await?
-            }
-            CachingStrategy::Truncate => {
-                self.update_last_modified_times(tables).await?;
-                self.delete_query_cache_entries(tables).await?
-            }
-            CachingStrategy::Memory(_) => {
-                self.update_last_modified_times(tables).await?;
-                clear_memory_query_cache(tables)?;
-            }
-        };
-        Ok(())
-    }
-
-    /// Update the cache tables for the given list of tables, using the current [CachingStrategy],
-    /// under the assumption that the tables in the given list have all just been dropped.
-    async fn clear_cache_for_dropped_tables(&self, tables: &[&str]) -> Result<(), DbError> {
-        if let CachingStrategy::Memory(_) = self.get_caching_strategy() {
-            self.update_last_modified_times(tables).await?;
-            clear_memory_query_cache(&tables)?;
-        } else {
-            // Do not clear the cache if the dropped tables include the cache tables themselves:
-            if !tables
-                .iter()
-                .any(|table| [QUERY_CACHE_TABLE, TABLE_CACHE_TABLE].contains(table))
-            {
-                match self.get_caching_strategy() {
-                    CachingStrategy::Memory(_) => unreachable!(),
-                    CachingStrategy::None => (),
-                    CachingStrategy::TruncateAll => {
-                        self.update_last_modified_times(tables).await?;
-                        self.delete_query_cache_entries(&[]).await?;
-                    }
-                    CachingStrategy::Trigger | CachingStrategy::Truncate => {
-                        self.update_last_modified_times(tables).await?;
-                        self.delete_query_cache_entries(tables).await?;
-                    }
-                }
-            }
-        }
-        // Update the meta-cache to remove any entries associated with tables that no longer exist:
-        let mut meta_cache = get_meta_cache()?;
-        for table in tables {
-            if *table == QUERY_CACHE_TABLE {
-                meta_cache.remove(QUERY_CACHE_TABLE);
-            } else if *table == TABLE_CACHE_TABLE {
-                meta_cache.remove(TABLE_CACHE_TABLE);
-            } else {
-                meta_cache.remove(&format!("{table}_triggers"));
-                meta_cache.remove(&format!("{table}_VIEW"));
-                meta_cache.remove(&format!("{table}_TABLE"));
-            }
-        }
-        Ok(())
-    }
-
-    /// Update the last verified time of the query cache entry identified by the triple:
-    /// (tables, statement, params).
-    async fn update_last_verified(
-        &self,
-        tables: &[&str],
-        statement: &str,
-        params: &DbParams,
-    ) -> Result<(), DbError> {
-        match self.get_caching_strategy() {
-            CachingStrategy::Memory(_) => {
-                let mut cache = get_memory_query_cache()?;
-                let epoch_now = SystemTime::now()
-                    .duration_since(UNIX_EPOCH)
-                    .map_err(|err| {
-                        DbError::DataError(format!("Error getting epoch time: {err}"))
-                    })?;
-                let mem_key = MemoryQueryCacheKey {
-                    tables: format!(
-                        "[{}]",
-                        tables
-                            .iter()
-                            .map(|table| format!("\"{table}\""))
-                            .collect::<Vec<_>>()
-                            .join(", ")
-                    ),
-                    statement: statement.to_string(),
-                    parameters: format!("{params:?}"),
-                };
-                match cache.get_mut(&mem_key) {
-                    Some(value) => value.last_verified = epoch_now.as_millis(),
-                    None => (),
-                };
-            }
-            _ => match self.table_exists(QUERY_CACHE_TABLE).await? {
-                true => {
-                    let tables_param = format!(
-                        "[{}]",
-                        tables
-                            .iter()
-                            .map(|table| format!("\"{table}\""))
-                            .collect::<Vec<_>>()
-                            .join(", ")
-                    );
-                    self.execute(
-                        &format!(
-                            r#"UPDATE "{QUERY_CACHE_TABLE}"
-                               SET "last_verified" = {ts}
-                               WHERE "tables" = {p}1
-                               AND "statement" = {p}2
-                               AND "parameters" = {p}3"#,
-                            p = self.kind().param_prefix(),
-                            ts = self.kind().get_epoch_time_sql(),
-                        ),
-                        &[
-                            &format!("[{tables_param}]"),
-                            statement,
-                            &format!("{params:?}"),
-                        ],
-                    )
-                    .await?;
-                }
-                false => (),
-            },
-        };
-        Ok(())
-    }
-
-    /// Update the last modified times of each of the given tables in the table cache.
-    async fn update_last_modified_times(&self, tables: &[&str]) -> Result<(), DbError> {
-        match self.get_caching_strategy() {
-            CachingStrategy::Memory(_) => {
-                let mut cache = get_memory_table_cache()?;
-                let epoch_now = SystemTime::now()
-                    .duration_since(UNIX_EPOCH)
-                    .map_err(|err| {
-                        DbError::DataError(format!("Error getting epoch time: {err}"))
-                    })?;
-                for table in tables {
-                    cache.insert(table.to_string(), epoch_now.as_millis());
-                }
-            }
-            _ => {
-                if self.table_exists(TABLE_CACHE_TABLE).await? {
-                    for table in tables {
-                        let sql = format!(
-                            r#"INSERT INTO "{TABLE_CACHE_TABLE}" ("table", "last_modified")
-                               VALUES ({prefix}1, {ts})
-                               ON CONFLICT ("table") DO UPDATE SET "last_modified" = {ts}"#,
-                            prefix = self.kind().param_prefix(),
-                            ts = self.kind().get_epoch_time_sql(),
-                        );
-                        self.execute_no_cache_clean(&sql, params![table]).await?;
-                    }
-                }
-            }
-        };
-        Ok(())
-    }
-
-    /// Delete the entries for the tables in the given list (independently of the current
-    /// caching strategy) from the query cache table, if it exists. If the given list is empty,
-    /// clear the entire query cache table.
-    async fn delete_query_cache_entries(&self, tables: &[&str]) -> Result<(), DbError> {
-        if self.table_exists(QUERY_CACHE_TABLE).await? {
-            if tables.is_empty() {
-                self.execute_no_cache_clean(&format!(r#"DELETE FROM "{QUERY_CACHE_TABLE}""#), ())
-                    .await?;
-            } else {
-                for table in tables {
-                    let table_param = format!(r#"%"{table}"%"#);
-                    self.execute_no_cache_clean(
-                        &format!(
-                            r#"DELETE FROM "{QUERY_CACHE_TABLE}" WHERE "tables" LIKE {}1"#,
-                            self.kind().param_prefix()
-                        ),
-                        &[table_param],
-                    )
-                    .await?;
-                }
-            }
-        }
-        Ok(())
-    }
 
     /// Get the SQL code that is used to define the given view.
     async fn get_view_sql(&self, view: &str) -> Result<String, DbError> {
@@ -514,358 +89,6 @@ pub trait DbQuery {
             }
         };
         Ok(view_sql)
-    }
-
-    /// Uses the current caching strategy to clear the query cache for any of the given tables
-    /// that (a) are views and (b) have source tables that have been modified more recently than
-    /// the view. This function works both with database and memory cache strategies.
-    async fn update_cached_views(&self, tables: &[&str]) -> Result<(), DbError> {
-        let views = self.which_are_views(tables).await?;
-        match self.get_caching_strategy() {
-            CachingStrategy::Memory(_) => {
-                for view in &views {
-                    let last_verified = {
-                        let mut last_verified = 0;
-                        for (key, value) in get_memory_query_cache()?.iter() {
-                            if key.tables.contains(view) && value.last_verified > last_verified {
-                                last_verified = value.last_verified;
-                            }
-                        }
-                        last_verified
-                    };
-                    let view_sql = self.get_view_sql(&view).await?;
-                    let view_tables = get_view_tables(&view_sql)?;
-                    let last_modified = {
-                        let mut latest_last_modified = 0;
-                        for view_table in &view_tables {
-                            match get_memory_table_cache()?.get(view_table) {
-                                Some(lm) if *lm > latest_last_modified => {
-                                    latest_last_modified = *lm;
-                                }
-                                _ => (),
-                            };
-                        }
-                        latest_last_modified
-                    };
-                    if last_modified >= last_verified {
-                        clear_memory_query_cache(&[view])?;
-                    }
-                }
-            }
-            _ => {
-                for view in &views {
-                    let last_verified = self.last_verified(&view).await?;
-                    let view_sql = self.get_view_sql(&view).await?;
-                    let view_tables = get_view_tables(&view_sql)?;
-                    let last_modified = self
-                        .get_latest_last_modified(
-                            &view_tables.iter().map(|t| t.as_str()).collect::<Vec<_>>(),
-                        )
-                        .await?;
-                    if last_modified >= last_verified {
-                        self.delete_query_cache_entries(&[&view]).await?;
-                    }
-                }
-            }
-        };
-        Ok(())
-    }
-
-    /// Returns the latest of the last modified times of the given tables in the table cache.
-    async fn get_latest_last_modified(&self, tables: &[&str]) -> Result<u64, DbError> {
-        match self.table_exists(TABLE_CACHE_TABLE).await? {
-            true => {
-                let prefix = self.kind().param_prefix().to_string();
-                let mut placeholders = vec![];
-                let mut parameters = vec![];
-                for (i, table) in tables.iter().enumerate() {
-                    let i = i + 1;
-                    placeholders.push(format!("{prefix}{i}"));
-                    parameters.push(DbValue::from(*table));
-                }
-                let placeholders = placeholders.join(",");
-
-                let sql = format!(
-                    r#"SELECT "last_modified"
-                       FROM "{TABLE_CACHE_TABLE}"
-                       WHERE "table" IN ({placeholders})
-                       ORDER BY "last_modified" DESC
-                       LIMIT 1"#,
-                );
-                let rows = self.query_no_cache_clean(&sql, parameters).await?;
-                match rows.len() {
-                    0 => Ok(0),
-                    1 => {
-                        let row = &rows[0];
-                        match row.get("last_modified") {
-                            Some(last_modified) => Ok(last_modified.try_into()?),
-                            None => Err(DbError::DataError(format!(
-                                "No field 'last_modified' in row {row:?}"
-                            ))),
-                        }
-                    }
-                    too_many => Err(DbError::DataError(format!(
-                        "Too many rows returned: {too_many} from table {TABLE_CACHE_TABLE}"
-                    ))),
-                }
-            }
-            false => Ok(0),
-        }
-    }
-
-    /// Gets the last time the given table was modified, as read from the table cache table.
-    /// If there is no entry for the table in the table cache, or if the table cache does not
-    /// exist, returns 0.
-    async fn last_modified(&self, table: &str) -> Result<u64, DbError> {
-        self.get_latest_last_modified(&[table]).await
-    }
-
-    /// Gets the last time that the given table was verified, as read from the query cache table.
-    /// If there is no entry involving the given table in the query cache, or if the query cache
-    /// table doesn't exist, returns 0.
-    async fn last_verified(&self, table: &str) -> Result<u64, DbError> {
-        match self.table_exists(QUERY_CACHE_TABLE).await? {
-            true => {
-                let sql = format!(
-                    r#"SELECT MAX("last_verified") AS "last_verified"
-                       FROM "{QUERY_CACHE_TABLE}"
-                       WHERE "tables" LIKE {p}1"#,
-                    p = self.kind().param_prefix(),
-                );
-                let table_param = format!(r#"%"{table}"%"#);
-                let rows = self.query_no_cache_clean(&sql, &[&table_param]).await?;
-                match rows.first() {
-                    Some(row) => match row.get("last_verified") {
-                        Some(value) if value == DbValue::Null => Ok(0),
-                        Some(value) => Ok(value.try_into()?),
-                        None => Err(DbError::DataError(format!(
-                            "No 'last_verified' found in row: {row:?}"
-                        ))),
-                    },
-                    None => Ok(0),
-                }
-            }
-            false => Ok(0),
-        }
-    }
-
-    /// Execute the given SQL command with the given parameters, returning a vector of rows.
-    /// If the result of the command exists in the query cache for the given tables, get the
-    /// value from there instead of from the tables themselves, in accordance with the given
-    /// [CachingStrategy].
-    async fn cache(
-        &self,
-        sql: &str,
-        params: impl IntoDbParams + Send + Copy + Sync,
-    ) -> Result<DbRows, DbError> {
-        match self.get_caching_strategy() {
-            CachingStrategy::None => self.cache_tables(&[], sql, params).await,
-            _ => {
-                let tables_read = get_accessed_tables(sql)?;
-                let tables_read: Vec<_> = tables_read.iter().map(|s| s.as_str()).collect();
-                match tables_read.is_empty() {
-                    false => self.cache_tables(&tables_read, sql, params).await,
-                    true => Err(DbError::InputError(format!(
-                        "No tables are read from in SQL: {sql}"
-                    ))),
-                }
-            }
-        }
-    }
-
-    /// Similar to [DbQuery::cache()]. This version accepts an explicit list of tables, which
-    /// must correspond to the tables queried from in the given SQL command(s).
-    async fn cache_tables(
-        &self,
-        tables: &[&str],
-        sql: &str,
-        params: impl IntoDbParams + Send + Copy + Sync,
-    ) -> Result<DbRows, DbError> {
-        let db_cache = async |tables: &[&str],
-                              sql: &str,
-                              params: &DbParams|
-               -> Result<DbRows, DbError> {
-            // Look in the cache to see if there is an entry corresponding to the given SQL
-            // string for the given tables and parameters. If so, return the data from the
-            // cache, otherwise execute the given SQL statement on the actualy specified
-            // tables.
-            let prefix = self.kind().param_prefix().to_string();
-            let cache_sql = format!(
-                r#"SELECT {prefix}1||rtrim(ltrim("value", '['), ']')||{prefix}2 AS "value"
-                       FROM "{QUERY_CACHE_TABLE}"
-                       WHERE "tables" = {prefix}3
-                       AND "statement" = {prefix}4
-                       AND "parameters" = {prefix}5
-                       LIMIT 1"#,
-            );
-            let tables_param = format!(
-                "[{}]",
-                tables
-                    .iter()
-                    .map(|table| format!("\"{table}\""))
-                    .collect::<Vec<_>>()
-                    .join(", ")
-            );
-            let params_param = match params {
-                DbParams::None => "[]".to_string(),
-                DbParams::Positional(params) => {
-                    let params = params.iter().map(|p| p.into()).collect::<Vec<String>>();
-                    format!("[{}]", params.join(", "))
-                }
-            };
-            let cache_params = &["[", "]", &tables_param, sql, &params_param];
-
-            let strings = {
-                let rows = self.query_no_cache_clean(&cache_sql, cache_params).await?;
-                let strings = rows
-                    .iter()
-                    .map(|row| match row.values().nth(0) {
-                        Some(value) => Ok(value.into()),
-                        None => Err(DbError::DataError("Empty row".to_owned())),
-                    })
-                    .collect::<Vec<_>>();
-                let strings: Result<Vec<String>, DbError> = strings.into_iter().collect();
-                strings?
-            };
-            match strings.first() {
-                Some(values) => {
-                    let db_rows: Vec<DbRow> = serde_json::from_str(values).map_err(|err| {
-                        DbError::DataError(format!("Error serializing values '{values}': {err}"))
-                    })?;
-                    // Only views need to be verified every time they are accessed. Tables
-                    // do not because they do not have any dependencies.
-                    if self.which_are_views(tables).await?.len() > 0 {
-                        self.update_last_verified(tables, sql, &params).await?;
-                    }
-                    Ok(DbRows { content: db_rows })
-                }
-                None => {
-                    let db_rows = self.query_no_cache_clean(sql, params).await?;
-                    let rows_as_string = {
-                        let mut rows_as_string = vec![];
-                        for db_row in db_rows.iter() {
-                            let db_row = serde_json::to_string(db_row).map_err(|err| {
-                                DbError::DataError(format!("Invalid data ({err}): {db_row:?}"))
-                            })?;
-                            rows_as_string.push(db_row);
-                        }
-                        format!("[{}]", rows_as_string.join(", "))
-                    };
-                    let insert_sql = format!(
-                        r#"INSERT INTO "{QUERY_CACHE_TABLE}"
-                           ("tables", "statement", "parameters", "value")
-                           VALUES ({prefix}1, {prefix}2, {prefix}3, {prefix}4)"#,
-                        prefix = self.kind().param_prefix(),
-                    );
-                    let insert_params = [&tables_param, sql, &params_param, &rows_as_string];
-                    self.execute_no_cache_clean(&insert_sql, &insert_params)
-                        .await?;
-                    Ok(db_rows)
-                }
-            }
-        };
-
-        let mem_cache = async |tables: &[&str],
-                               sql: &str,
-                               params: &DbParams,
-                               cache_size: usize|
-               -> Result<DbRows, DbError> {
-            let params = &params.into_db_params();
-            let mem_key = MemoryQueryCacheKey {
-                tables: format!(
-                    "[{}]",
-                    tables
-                        .iter()
-                        .map(|table| format!("\"{table}\""))
-                        .collect::<Vec<_>>()
-                        .join(", ")
-                ),
-                statement: sql.to_string(),
-                parameters: format!("{params:?}"),
-            };
-            let cached_rows = {
-                let cache = get_memory_query_cache()?;
-                match cache.get(&mem_key) {
-                    Some(mem_value) => Some(mem_value.content.to_vec()),
-                    None => None,
-                }
-            };
-            match cached_rows {
-                Some(db_rows) => {
-                    // Only views need to be verified every time they are accessed. Tables
-                    // do not because they do not have any dependencies.
-                    if self.which_are_views(tables).await?.len() > 0 {
-                        self.update_last_verified(tables, sql, &params).await?;
-                    }
-                    Ok(DbRows { content: db_rows })
-                }
-                None => {
-                    let db_rows = self.query_no_cache_clean(sql, params).await?;
-                    let mut cache = get_memory_query_cache()?;
-                    // If the number of entries exceeds the allowed cache size, remove the oldest
-                    // keys first.
-                    // TODO: We may want to do something smarter here. E.g., we could record the
-                    // length of time a query takes and/or the number of times it was requested
-                    // in order to determine which entries to delete (and/or the size of the
-                    // result).
-                    while cache.len() > cache_size {
-                        cache.shift_remove_index(0);
-                    }
-                    cache.insert(
-                        mem_key,
-                        MemoryQueryCacheValue {
-                            content: db_rows.to_vec(),
-                            last_verified: SystemTime::now()
-                                .duration_since(UNIX_EPOCH)
-                                .map_err(|err| {
-                                    DbError::DataError(format!("Error getting epoch time: {err}"))
-                                })?
-                                .as_millis(),
-                        },
-                    );
-                    Ok(db_rows)
-                }
-            }
-        };
-
-        match self.get_caching_strategy() {
-            CachingStrategy::None => {
-                let rows = self.query_no_cache_clean(sql, params).await?;
-                Ok(rows)
-            }
-            CachingStrategy::TruncateAll | CachingStrategy::Truncate => {
-                self.ensure_cache_tables_exist().await?;
-                self.update_cached_views(tables).await?;
-                let rows = db_cache(tables, sql, &params.into_db_params()).await?;
-                Ok(rows)
-            }
-            CachingStrategy::Trigger => {
-                let views = self
-                    .which_are_views(tables)
-                    .await?
-                    .into_iter()
-                    .collect::<HashSet<_>>();
-                for table in tables
-                    .into_iter()
-                    .cloned()
-                    .collect::<HashSet<_>>()
-                    .difference(&views.iter().map(|v| v.as_str()).collect::<HashSet<_>>())
-                    .collect::<Vec<_>>()
-                {
-                    self.ensure_caching_triggers_exist_for_table(&table).await?;
-                }
-                for view in &views {
-                    self.ensure_caching_triggers_exist_for_view(view).await?;
-                }
-                let rows = db_cache(&tables, sql, &params.into_db_params()).await?;
-                Ok(rows)
-            }
-            CachingStrategy::Memory(cache_size) => {
-                self.update_cached_views(tables).await?;
-                let rows = mem_cache(tables, sql, &params.into_db_params(), cache_size).await?;
-                Ok(rows)
-            }
-        }
     }
 
     /// Given a table, return a [ColumnMap] from column names to column SQL types.
@@ -946,13 +169,11 @@ pub trait DbQuery {
     fn execute_batch(&self, sql: &str) -> impl Future<Output = Result<(), DbError>> + Send;
 
     /// Execute the given SQL command, with the given parameters, returning a vector of rows.
-    async fn query(&self, sql: &str, params: impl IntoDbParams + Send) -> Result<DbRows, DbError> {
-        let rows = self.query_no_cache_clean(sql, params).await?;
-        if self.get_cache_aware_query() {
-            self.clear_cache_for_affected_tables(sql).await?;
-        }
-        Ok(rows)
-    }
+    fn query(
+        &self,
+        sql: &str,
+        params: impl IntoDbParams + Send,
+    ) -> impl Future<Output = Result<DbRows, DbError>> + Send;
 
     /// Execute the given SQL command using the given parameters, returning a vector of rows,
     /// without updating the cache, regardless of whether the cache-aware-query option
@@ -1130,45 +351,8 @@ pub trait DbQuery {
     /// <https://www.postgresql.org/docs/current/sql-droptable.html>), if the dropped table,
     /// say table1, appears in a foreign key constraint for another table, say table2, then
     /// table2's foreign constraint will be removed, but table2 will not be dropped.
-    async fn drop_table(&self, table: &str) -> Result<(), DbError> {
-        let table = validate_table_name(table)?;
-        // Drop the table:
-        match self.kind() {
-            DbKind::PostgreSQL => {
-                self.execute_no_cache_clean(
-                    &format!(r#"DROP TABLE IF EXISTS "{table}" CASCADE"#),
-                    (),
-                )
-                .await?
-            }
-            DbKind::SQLite => {
-                self.execute_no_cache_clean(&format!(r#"DROP TABLE IF EXISTS "{table}""#), ())
-                    .await?
-            }
-        };
-
-        // Delete dirty entries from the cache in accordance with our caching strategy:
-        self.clear_cache_for_dropped_tables(&[&table]).await?;
-        Ok(())
-    }
+    fn drop_table(&self, table: &str) -> impl Future<Output = Result<(), DbError>> + Send;
 
     /// Drop the given view from the database.
-    async fn drop_view(&self, view: &str) -> Result<(), DbError> {
-        let view = validate_table_name(view)?;
-        // Drop the view:
-        match self.kind() {
-            DbKind::PostgreSQL => {
-                self.execute_no_cache_clean(&format!(r#"DROP VIEW IF EXISTS "{view}" CASCADE"#), ())
-                    .await?
-            }
-            DbKind::SQLite => {
-                self.execute_no_cache_clean(&format!(r#"DROP VIEW IF EXISTS "{view}""#), ())
-                    .await?
-            }
-        };
-
-        // Delete dirty entries from the cache in accordance with our caching strategy:
-        self.clear_cache_for_dropped_tables(&[&view]).await?;
-        Ok(())
-    }
+    fn drop_view(&self, view: &str) -> impl Future<Output = Result<(), DbError>> + Send;
 }
