+++ src/shared.rs
@@ -1,5 +1,6 @@
 use crate::{
-    core::{DbError, DbQuery},
+    cache::DbCache,
+    core::DbError,
     db_kind::DbKind,
     db_value::{DbRows, DbValue, IntoDbRows},
     parse::validate_table_name,
@@ -130,7 +131,7 @@ ON CONFLICT ({constraint_clause}) DO UPDATE SET {set_clause}{returning_clause}"#
 /// used to edit the table, do not use more than max_params bound parameters at a time. If more
 /// than max_params are required, multiple SQL statements will be generated.
 pub(crate) async fn edit(
-    pool: &(impl DbQuery + Sync),
+    pool: &(impl DbCache + Sync),
     edit_type: &EditType,
     max_params: &usize,
     table: &str,
